+++ src/utility/getRandomDirection.js
@@ -1,0 +1,6 @@
+const getRandomDirection = function () {
+    const scenesArray = ['woodScene', 'riverScene', 'gardenScene'];
+    return scenesArray[Math.floor(Math.random() * Math.floor(scenesArray.length))]
+};
+
+export {getRandomDirection}
