+++ src/utility/getRandomKey.js
@@ -1,0 +1,5 @@
+const getRandomKey = function (targetObject) {
+    return targetObject[Object.keys(targetObject)[Math.floor(Math.random() * Math.floor(Object.keys(targetObject).length))]];
+};
+
+export {getRandomKey}
