+++ src/utility/killChildren.js
@@ -1,0 +1,7 @@
+const killChildren = function (parent) {
+    while (parent.firstChild) {
+        parent.removeChild(parent.firstChild);
+    }
+};
+
+export {killChildren}
