+++ src/utility/killPlayer.js
@@ -1,0 +1,13 @@
+import {playerStatus} from "../content/playerStatus";
+import {renderScene} from "./renderScene";
+
+const killPlayer = function () {
+    if (playerStatus.health <= 0 ||
+        playerStatus.fatigue >= 15 ||
+        playerStatus.hunger >= 15 ||
+        playerStatus.thirst >= 10) {
+        renderScene('youDiedScene');
+    }
+};
+
+export {killPlayer}
